asdasd

+++ src/misc/arrange_data_for_fmri_preproc.py
@@ -45,11 +45,12 @@ def arrange_data(subjects, trs):
         for bold_fol in bold_fols:
             fmri_files = glob.glob(op.join(bold_fol, '*_rest_reorient_skip_faln_mc.nii.gz'))
             if len(fmri_files) == 1:
-                target_file = op.join(sub_fmri_fol, '{}_bld{}_rest_reorient.nii.gz'.format(sub, utils.namebase(bold_fol)))
                 sub_fmri_fol = utils.make_dir(op.join(sub_root_bold_fol, utils.namebase(bold_fol)))
+                target_file = op.join(sub_fmri_fol, '{}_bld{}_rest_reorient.nii.gz'.format(sub, utils.namebase(bold_fol)))
                 # make_link(fmri_files[0], target_file)
                 remove_link(target_file)
-                shutil.copy(fmri_files[0], target_file)
+                if not op.isfile(target_file):
+                    shutil.copy(fmri_files[0], target_file)
             elif len(fmri_files) == 0:
                 print('No *_rest_reorient_skip_faln_mc.nii.gz files were found in {}!'.format(bold_fol))
                 continue
